refactor: migrate spinner calls to tui package and remove unused quality parameter in Goyabu provider
